server/requirements.txt
scikit-learn>=1.0
catboost>=1.0
python-dotenv>=0.21
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
